sagemaker/requirements.txt
absl-py==0.9.0
albumentations==0.4.5
antlr4-python3-runtime==4.7.2
astor==0.8.1
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
Cython==0.29.19
decorator==4.4.2
dlr==1.0
gast==0.2.2
gluoncv==0.7.0
google-pasta==0.2.0
graphviz==0.8.4
grpcio==1.29.0
h5py==2.10.0
idna==2.9
imageio==2.8.0
imgaug==0.2.6
importlib-metadata==1.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.2.1
mxnet==1.6.0
networkx==2.4
numpy==1.18.4
onnx==1.7.0
opencv-python-headless==4.2.0.34
opt-einsum==3.2.1
Pillow==7.1.2
pkg-resources==0.0.0
portalocker==1.7.0
protobuf==3.12.1
psutil==5.7.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.24.0
scikit-image==0.17.2
scipy==1.4.1
six==1.15.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.3
termcolor==1.1.0
tifffile==2020.6.3
tornado==6.0.4
tqdm==4.46.1
typing-extensions==3.7.4.2
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
xgboost==1.1.0
zipp==3.1.0
